+++ to-dolist/src/main/java/com/crud/todolist/entities/StatusTarefa.java
@@ -27,8 +27,7 @@ public enum StatusTarefa {
 		for(StatusTarefa atual:values()) {
 			if(atual.getValue() == id) {
 				return atual;
-			}
-			
+			} 			
 		}
 		return null;
 	}