+++ to-dolist/src/main/java/com/crud/todolist/repositories/TarefaRequest.java
@@ -1,6 +1,7 @@
 package com.crud.todolist.repositories;
 
+import jakarta.validation.constraints.NotBlank;
 
-public record TarefaRequest(Long id, String titulo, String descricao, String dataCriacao, Integer status) {
+public record TarefaRequest(Long id, @NotBlank String titulo, String descricao, String dataCriacao, Integer status) {
 
 }