+++ to-dolist/src/main/resources/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [v1.1] - 2025-01-23
 ### Acréscimo do status
-- Implementação do status da tarefa
+- Implementação do status da tarefa 
 
 ## [v1.0] - 2025-01-23
 ### Primeiro lançamento estável